In-progress

+++ engine/src/global.hpp
@@ -42,24 +42,24 @@ namespace wisdom
 
     constexpr int Num_Players = 2;
 
-    constexpr int8_t Num_Rows = 8;
-    constexpr int8_t Num_Columns = 8;
+    constexpr int Num_Rows = 8;
+    constexpr int Num_Columns = 8;
 
-    constexpr int8_t First_Row = 0;
-    constexpr int8_t First_Column = 0;
+    constexpr int First_Row = 0;
+    constexpr int First_Column = 0;
 
-    constexpr int8_t Last_Row = 7;
-    constexpr int8_t Last_Column = 7;
+    constexpr int Last_Row = 7;
+    constexpr int Last_Column = 7;
 
-    constexpr int8_t King_Column = 4;
-    constexpr int8_t King_Rook_Column = 7;
-    constexpr int8_t Queen_Rook_Column = 0;
+    constexpr int King_Column = 4;
+    constexpr int King_Rook_Column = 7;
+    constexpr int Queen_Rook_Column = 0;
 
-    constexpr int8_t Kingside_Castled_Rook_Column = 5;
-    constexpr int8_t Queenside_Castled_Rook_Column = 3;
+    constexpr int Kingside_Castled_Rook_Column = 5;
+    constexpr int Queenside_Castled_Rook_Column = 3;
 
-    constexpr int8_t Kingside_Castled_King_Column = 6;
-    constexpr int8_t Queenside_Castled_King_Column = 2;
+    constexpr int Kingside_Castled_King_Column = 6;
+    constexpr int Queenside_Castled_King_Column = 2;
 
     // Infinity score.
     constexpr int Infinity = 65536;
